+++ src/shared/types/index.ts
@@ -2,3 +2,4 @@ export * from './entities';
 export * from './connections';
 export * from './embeddings';
 export * from './clusters';
+export * from './resources';